+++ core/src/main/res/layout/base_testpress_webview_container_layout.xml
@@ -14,11 +14,4 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
 
-    <FrameLayout
-        android:layout_below="@id/appbar"
-        android:id="@+id/empty_view_container"
-        android:visibility="gone"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
-
 </RelativeLayout>
